fix(app): rename loadingscreen to loading
